+++ src/Hakamiq.Cso.Core/Formats/Cso/CompressionMethod.cs
@@ -4,4 +4,5 @@ public enum CompressionMethod
 {
     Store = 0,
     RawDeflate = 1,
+    ZopfliDeflate = 2,
 }